tests/wpt/web-platform-tests/webmessaging/with-options/null-arg-two.tentative.html

<!doctype html>
<title>null arg two interpreted as null string</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id=log></div>
<script>
test(function(t) {
  assert_throws(new SyntaxError, () => postMessage('', null));
});
</script>
